backend/requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.2
PyJWT==2.11.0
python-dotenv==1.0.0